Moved all routines that use BandedMatrix structure to hacks
